+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - improve stake epic error handling
 - hide stake button after staking started in Staking overview page
 - able to overwrite WEBSOCKET_LATENCY_PERIOD from runtime via cli args
+- able to switch network with different remote api settings
 
 # v0.6.1 (2019.03.12)